debt - inline (scary) stats.ts into diagnostics service where it is used
